refactor : Post domain 경로 수정
1. Post 도메인 경로 분리
- domain, request, response
