+++ threads/src/main/java/com/imwoo/threads/model/post/request/PostCreateRequest.java
@@ -1,4 +1,4 @@
-package com.imwoo.threads.model;
+package com.imwoo.threads.model.post.request;
 
 public record PostCreateRequest(String body) {
 }
