+++ threads/src/main/java/com/imwoo/threads/repository/PostEntityRepository.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.imwoo.threads.model.PostResponse;
 import com.imwoo.threads.model.entity.PostEntity;
+import com.imwoo.threads.model.post.response.PostResponse;
 
 public interface PostEntityRepository extends JpaRepository<PostEntity, Long> {
