+++ threads/src/main/java/com/imwoo/threads/service/PostService.java
@@ -6,10 +6,10 @@ import org.springframework.stereotype.Service;
 
 import com.imwoo.threads.exception.post.PostCreatedFailureException;
 import com.imwoo.threads.exception.post.PostNotFoundException;
-import com.imwoo.threads.model.PostCreateRequest;
-import com.imwoo.threads.model.PostResponse;
-import com.imwoo.threads.model.PostUpdateRequest;
 import com.imwoo.threads.model.entity.PostEntity;
+import com.imwoo.threads.model.post.request.PostCreateRequest;
+import com.imwoo.threads.model.post.request.PostUpdateRequest;
+import com.imwoo.threads.model.post.response.PostResponse;
 import com.imwoo.threads.repository.PostEntityRepository;
 
 import lombok.RequiredArgsConstructor;
